Envio final

+++ src/main/java/com/broadfactor/desafio/controller/UserController.java
@@ -6,6 +6,7 @@ import com.broadfactor.desafio.model.UserDetailsResponse;
 import com.broadfactor.desafio.model.UserEntity;
 import com.broadfactor.desafio.repository.UserRepository;
 import com.broadfactor.desafio.service.UserService;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,14 +18,22 @@ import java.util.UUID;
 
 @RestController
 @RequestMapping("/user")
+@SecurityRequirement(name = "bearer-key")
 public class UserController {
     @Autowired
     private UserService userService;
     @Autowired
     private UserRepository userRepository;
     @PutMapping("/{id}")
-    public Object update(@PathVariable UUID id, @RequestBody UserDTO dados){
-        return userService.update(id, dados);
+    public ResponseEntity<UserDetailsDTO> update(@PathVariable UUID id, @RequestBody UserDTO dados){
+       var valida = userRepository.getReferenceById(id);
+       if (valida != null){
+            var updated = userService.update(id, dados);
+            if (updated != null){
+                return ResponseEntity.ok(new UserDetailsDTO(updated.getId(), updated.getName(), updated.getEmail(), updated.getCompany()));
+            }
+       }
+        return ResponseEntity.badRequest().build();
     }
 
     @GetMapping
@@ -41,7 +50,10 @@ public class UserController {
     @DeleteMapping("/{id}")
     public ResponseEntity delete(@PathVariable UUID id){
         var user = userRepository.getReferenceById(id);
-        userRepository.delete(user);
-        return ResponseEntity.noContent().build();
+        if (user != null) {
+            userRepository.delete(user);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
